move datastore functions into ip_pool file
